Change HID_::SendReport "id" argument from 16 to 8bits.

+++ src/HID/HID.cpp
@@ -145,7 +145,7 @@ int HID_::SetFeature(uint16_t id, const void* data, int len)
     return reportCount;
 }
 
-int HID_::SendReport(uint16_t id, const void* data, int len)
+int HID_::SendReport(uint8_t id, const void* data, int len)
 {
     auto ret = USB_Send(pluggedEndpoint, &id, 1);
     if (ret < 0) return ret;
